refactor: transport ディレクトリを <name>http にリネームし dir名=package名に統一
HTTP 層は api 型境界のため別パッケージ <name>http のままだが、ディレクトリ名が
transport で package 名と不一致だった。ディレクトリを <name>http にリネームし
Go 慣習（dir名=package名）に合わせる。main.go/router.go の import エイリアスを除去し、
.go-arch-lint.yml の in: パスを更新。package 宣言・型名・境界強制は不変。
