+++ cmd/api/main.go
@@ -13,17 +13,17 @@ import (
 	"stock_backend/internal/app/config"
 	"stock_backend/internal/app/router"
 	"stock_backend/internal/feature/auth"
-	authhttp "stock_backend/internal/feature/auth/transport"
+	"stock_backend/internal/feature/auth/authhttp"
 	"stock_backend/internal/feature/candles"
-	candleshttp "stock_backend/internal/feature/candles/transport"
+	"stock_backend/internal/feature/candles/candleshttp"
 	"stock_backend/internal/feature/logodetection"
 	"stock_backend/internal/feature/logodetection/gemini"
-	logodetectionhttp "stock_backend/internal/feature/logodetection/transport"
+	"stock_backend/internal/feature/logodetection/logodetectionhttp"
 	"stock_backend/internal/feature/logodetection/vision"
 	"stock_backend/internal/feature/symbollist"
-	symbollisthttp "stock_backend/internal/feature/symbollist/transport"
+	"stock_backend/internal/feature/symbollist/symbollisthttp"
 	"stock_backend/internal/feature/watchlist"
-	watchlisthttp "stock_backend/internal/feature/watchlist/transport"
+	"stock_backend/internal/feature/watchlist/watchlisthttp"
 	infradb "stock_backend/internal/platform/db"
 	"stock_backend/internal/platform/httpratelimit"
 	jwtmw "stock_backend/internal/platform/jwt"
