+++ internal/app/router/router.go
@@ -7,11 +7,11 @@ import (
 	"github.com/gin-contrib/cors"
 	"github.com/gin-gonic/gin"
 
-	authhttp "stock_backend/internal/feature/auth/transport"
-	candleshttp "stock_backend/internal/feature/candles/transport"
-	logodetectionhttp "stock_backend/internal/feature/logodetection/transport"
-	symbollisthttp "stock_backend/internal/feature/symbollist/transport"
-	watchlisthttp "stock_backend/internal/feature/watchlist/transport"
+	"stock_backend/internal/feature/auth/authhttp"
+	"stock_backend/internal/feature/candles/candleshttp"
+	"stock_backend/internal/feature/logodetection/logodetectionhttp"
+	"stock_backend/internal/feature/symbollist/symbollisthttp"
+	"stock_backend/internal/feature/watchlist/watchlisthttp"
 	csrfmw "stock_backend/internal/platform/csrf"
 	handler "stock_backend/internal/platform/handler"
 	"stock_backend/internal/platform/httpratelimit"
