+++ internal/feature/logodetection/logodetectionhttp/handler.go
