+++ internal/feature/symbollist/symbollisthttp/handler.go
